scraper/requirements.txt
playwright==1.44.0
beautifulsoup4==4.12.3
httpx==0.27.0
supabase==2.5.3
python-dotenv==1.0.1
lxml==5.2.2
